refactor App to components directory logo to assets directory
fix import paths so everything compiles properly
