+++ src/components/App/index.js
@@ -1,6 +1,6 @@
 import React from 'react';
-import logo from './logo.svg';
-import './App.css';
+import logo from '../../assets/logo.svg';
+import './index.css';
 
 function App() {
   return (
